reorganise source code in preparation for cmake build
