+++ example/Counter.test.cpp
@@ -1,10 +1,13 @@
 #include <gtest/gtest.h>
 using namespace testing;
 
-#include "gtestverilog/gtestverilog.h"
+#include "googletest-verilog/googletest-verilog.h"
 using namespace gtestverilog;
 
-#include "gtestverilog/CounterTestBench.h"
+// JK - disable, until we get automated test bench generation working again
+#if 0
+
+#include "googletest-verilog-example/CounterTestBench.h"
 using namespace countertestbench;
 
 namespace {
@@ -89,3 +92,5 @@ TEST_F(Counter, ShouldHandleLargeTraces) {
 
     ASSERT_THAT(testBench.trace, MatchesTrace(traceExpected));
 }
+
+#endif
