--- gtestverilog/gtestverilog.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "lib/MatchesTrace.h"
-#include "lib/TestBench.h"
-#include "lib/TraceBuilder.h"
