+++ include/googletest-verilog/ConsoleColour.h
