+++ include/googletest-verilog/Trace.h
