+++ include/googletest-verilog/TraceBuilder.h
