+++ include/googletest-verilog/googletest-verilog.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "MatchesTrace.h"
+#include "TestBench.h"
+#include "TraceBuilder.h"
